requirements.txt
python-dotenv
django
requests